requirements.txt
# Engage AI Conversation Validator Requirements

# Core ML and NLP
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
spacy>=3.6.0
numpy>=1.24.0

# Data handling
pandas>=2.0.0
pyyaml>=6.0

# Async support
asyncio-subprocess>=0.1.0

# Install spacy model after pip install:
# python -m spacy download en_core_web_sm